Rewrite the Object.create example.
The previous version worked just the same even when deleting the line which had Object.create in it!
Also, this version seems simpler to me, as it does not assume knowledge of functions as constructors and the new keyword, etc.
This version however assumes (teaches?) knowledge about the context of inherited methods. It could of course be even simpler, if we just leave out the sayName method and focus on primitive inherited fields.