More illustrative example for `Proxy.apply` (#677)
IMO, the original example is somewhat lackluster.

In the existing example, the proxied `proxy1` function returns the same result as the original `sum` function; both return the sum of `arguments[0]` and `arguments[1]`. Th me, this seems pointless and does not do a proper job of showing how useful `Proxy` can be.

I think it would be best to show how `Proxy.apply` can be used to intercept and modify the return value and/or parameters passed to the target function.

Interested in everybody's thoughts. Thanks.