Issue #730 - Content Sectioning - header (#789)
* Issue #730 - Content Sectioning - header

Added example for the `<header>` element.

* Address review concerns

This commit corrects the following issues:

* Indentation fixed
* Removed use of background color and the border line.
  instead, the header image is used to fill the entire
  header background
* The Dancing Script font is now local
* Substantial rework of the CSS to optimize as much as
  possible
* Fixed problems with the layout breaking at narrow
  widths
* Shortened up the HTML by getting rid of the fake
  blog post and just having short body text instead
* Added attribution for the background image

* Changed tabs to spaces in header.css. Not sure how
they got there but they're fixed.

* Use a flexbox layout and calc() with vw units to set
the size of the header's font so that the text is
scaled between 1em and 3em depending on the window
width. Use a short, wide image as the backdrop. Change
the body to match other examples.

Moved font-face stuff to the end so it's not hogging
valuable space above the fold.

* Replaced a stray tab with two spaces, and switched to
using single quotes. Also outright removed quotes in
a couple of places since they were optional there
anyway.

* Forgot to un-escape `<header>` in the meta.json file.
Fixed.