fix(performance): send post to kuma for custom ie-load-event-end metric (#1001)
This will post a message to Kuma to set a custom mark at the same time as we send `performance.timing.loadEventEnd` for GA beaconing.

On the Kuma side I will create a new measure which will use `performance.timing. navigationStart` as the `startMark` and then the mark we are creating here as `endMark`. That should then give us the duration equal to `ie-load-complete`