Issue#946 open links in new tab (#1009)
I added two helper functions that are called inside the editor's `render` method.

`openLinksInNewTab` prevents the default click event on external links (all anchors with an `href` attribute that starts with `http`) and opens them in a new tab.

While I was at it, I stumbled upon the issue that relative links (like `<a href="#someanchor">`) also didn't work. I didn't find a ticket for that so I thought it would make sense to fix in the scope of this ticket.

`scrollToAnchors` prevents the default click event on relative links and scrolls the targeted anchor into view.

I hope that the code structure adheres to your quality standards.