Update <div> example to use cc0 image (#1088)
* Update <div> example to use cc0 image
Replaced the leopard.jpg image with a CC0-licensed one
and removed the comment with the attribution.

* Replace image in <figcaption> sample with CC0 one
Replaced the hamster (which was free but not expressly
CC0) with a known CC0 image. Updated the CSS and HTML
to use it and make it look right in the sample.

* Revert "Replace image in <figcaption> sample with CC0 one"
Accidentally includes the <div> sample. Will re-commit
soon.
This reverts commit 22350b93f1cb1c0534e1fa94c52f53db712929c2.