Provide valid and more semantic HTML with <audio> (#1236)
* Provide valid and more semantic HTML with <audio>

The old example used <label>, which while permitted in this context, the
value of `for` must be a labelable form element. In HTML 5.2, that list
is at https://www.w3.org/TR/html/sec-forms.html#labelable-element (W3C)
https://html.spec.whatwg.org/multipage/forms.html#category-label (TLS).
Because <audio> is not in that list, the previous example triggered
validation errors with both the W3C and WHATWG validators.

I have rewritten the example with a more accessible, semantic solution
using the <figure> element. The accompanying <figcaption> serves a more
semantic role than <label> in this context.

I've also updated the CSS for example, although it is kind of
superfluous.

* Rewrite CSS to purposefully present figure