demonstration of mutation side-effects
As the first example that the visitor will see, it's important to emphasize that the target argument will get mutated.  Those details get described a little later in the document, but it's easily missed at a glance.  This update to the example intends to demonstrate that.  Please advise if better language could be used.