Added name attribute to select (#1453)
The description for the select contains the following text: "The above example shows typical <select> usage. It is given an id attribute to enable it to be associated with a <label> for accessibility purposes, as well as a name attribute to represent the name of the associated data point submitted to the server. "

The name attribute mentioned in the text is missing in the example.