Update intl-collator-prototype-compare.html
This example was incorrect. The name of the function is `firstAlphabetical` however, the conditional check and expectations were the inverse.

I added an example in english which shows one many reviewers will recognize as correct ('a' is first, not 'z').  My knowledge of other alphabets and collation rules are less good, but some searching and wikipedia seem to confirm that these are correct (and as I would have expected based on the incorrect comparator, simply reversed from the original examples)