Added explicit "true" to contenteditable attribute
From https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/contenteditable 

"This attribute is an enumerated one and not a Boolean one. This means that the explicit usage of one of the values true, false or the empty string is mandatory and that a shorthand like <label contenteditable>Example Label</label> is not allowed. The correct usage is <label contenteditable="true">Example Label</label>."