Significantly improve by adding an ALL CAPS word
This example will benefit significantly by adding a word in ALL CAPS so developers can see how an all caps word is affected.

Fortunately, the first word of the quote from Dickens affords this opportunity perfectly.  According to [GoodReads][https://www.goodreads.com/quotes/195229-london-michaelmas-term-lately-over-and-the-lord-chancellor-sitting], the word "London" should be in all caps, so this improvement is a win-win.