CSS live examples: color contrast improvements (#2848)
* css logical-properties: Set fg & bg color

For the demonstration elements with
background-color: yellow, also set explicit
foreground color. Prevents white-on-yellow
text, with the dark skin.

* css content-visibility: Set fg & bg color

For demonstration elements with background-color: set and text content,
also set explicit foreground color.

* CSS filter example: Add white container element

The default filter in the example is a drop-shadow, which will be
invisible in dark mode against the default black background. Add an
`.example-container` div with background-color set to `#fff`, so
that the effect is visible regardless of the selected theme.